nit renaming + validation of valSetCap chain spec

+++ mod/chain-spec/pkg/chain/data.go
@@ -166,7 +166,7 @@ type SpecData[
 	// ValidatorSetCap is the maximum number of validators that can be active
 	// for a given epoch
 	// Note: ValidatorSetCap must be smaller than ValidatorRegistryLimit.
-	ValidatorSetCap uint32 `mapstructure:"validator-set-cap-size"`
+	ValidatorSetCap uint64 `mapstructure:"validator-set-cap-size"`
 	// EVMInflationAddress is the address on the EVM which will receive the
 	// inflation amount of native EVM balance through a withdrawal every block.
 	EVMInflationAddress ExecutionAddressT `mapstructure:"evm-inflation-address"`
